Fix tests - folder assign - unassign
